Revert "Files renamed to be more precise, typo in comment fixed."
This reverts commit 11576c094200700b27615db635ab22a9d631f83b.
