+++ monkey/infection_monkey/transport/attack_telems/victim_host_telem.py
@@ -7,7 +7,7 @@ class VictimHostTelem(AttackTelem):
 
     def __init__(self, technique, status, machine, data=None):
         """
-        ATT&CK telemetry that parses and sends VictimHost's (remote machine's) data
+        ATT&CK telemetry that parses and sends VictimHost telemetry
         :param technique: Technique ID. E.g. T111
         :param status: int from ScanStatus Enum
         :param machine: VictimHost obj from model/host.py
