+++ monkey/monkey_island/cc/app.py
@@ -30,7 +30,7 @@ from cc.resources.telemetry_feed import TelemetryFeed
 from cc.resources.pba_file_download import PBAFileDownload
 from cc.services.config import ConfigService
 from cc.resources.pba_file_upload import FileUpload
-from cc.resources.attack_telem import AttackTelem
+from cc.resources.attack import Attack
 
 __author__ = 'Barak'
 
@@ -124,6 +124,6 @@ def init_app(mongo_url):
                      '/api/fileUpload/<string:file_type>?load=<string:filename>',
                      '/api/fileUpload/<string:file_type>?restore=<string:filename>')
     api.add_resource(RemoteRun, '/api/remote-monkey', '/api/remote-monkey/')
-    api.add_resource(AttackTelem, '/api/attack/<string:technique>')
+    api.add_resource(Attack, '/api/attack/<string:technique>')
 
     return app
