+++ monkey/monkey_island/cc/services/attck/__init__.py
